# openapi.yaml
openapi: 3.0.3
info:
  title: Movie Recap Pipeline API
  description: Multi-tenant backend for automated long-form video recap generation
  version: 1.0.0
  contact:
    name: API Support
    email: support@movierecap.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.movierecap.com/v1
    description: Production server
  - url: https://staging-api.movierecap.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Development server
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: Authenticate user and return tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
...
  /uploads/init:
    post:
      tags: [Uploads]
      summary: Initialize resumable upload
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadInit'
      responses:
        '200':
          description: Upload initialized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadSession'
  /uploads/{upload_id}/complete:
    post:
      tags: [Uploads]
      summary: Complete upload and trigger processing
      security:
        - bearerAuth: []
      parameters:
        - name: upload_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Upload completed, processing started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'

  # Job Management
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: bearer
        expires_in:
          type: integer
...
    UploadInit:
      type: object
      required: [project_id, file_type, file_size, filename]
      properties:
        project_id:
          type: string
          format: uuid
        file_type:
          type: string
          enum: [video, script]
        file_size:
          type: integer
        filename:
          type: string
        content_type:
          type: string
    UploadSession:
      type: object
      properties:
        upload_id:
          type: string
          format: uuid
        drive_folder_id:
          type: string
        upload_url:
          type: string
        expires_at:
          type: string
          format: date-time
    JobResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, processing, manual_review, completed, failed]
...